+++ src/icons/study-icon.tsx
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2011-2020 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+/**
+ * Study icon
+ */
+export const StudyIcon = createIcon(
+    <svg viewBox="0 0 20 20" fill="none">
+        <path
+            d="M3 12a1 1 0 01-1-1V5a3 3 0 013-3h4.382a1 1 0 01.894.553l.448.894a1 1 0 00.894.553H16a2 2 0 012
+2v5a1 1 0 01-1 1H3z"
+            fill="#B6D477"
+        />
+        <path
+            d="M2 11a5 5 0 015-5h9a2 2 0 012 2v7a2 2 0 01-2 2H4a2 2 0 01-2-2v-4z"
+            fill="#6C990F"
+        />
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M15.003 15.465a4 4 0 10-4-6.928c-1.914 1.104-3.435 4.05-2.33 5.964 1.005 1.741 4.827 1.831
+6.33.964zM12.827 12c.829 0 2.173-.42 2.173-1.25C15 9.923 14.156 9 13.327 9c-.828 0-2.015.922-2.015
+1.75 0 .83.687 1.25 1.515 1.25zm.938.79a1 1 0 00.484 1.94c.536-.133 1.226-1.282
+1.093-1.818-.123-.494-.906-.294-1.447-.155l-.13.033zm-2.028.645c.271.42.122 1.112-.298
+1.383-.42.27-1.111.122-1.382-.298-.271-.42-.697-1.336.027-1.803.723-.467 1.382.298 1.653.718z"
+            fill="#B6D477"
+        />
+    </svg>
+);